Split Azure DevOps summary reporter
